requirements.txt
fastapi
uvicorn
httpx
requests
supabase    
python-dotenv
gunicorn
passlib[bcrypt]
pydantic[email]